better remove

+++ src/app/open/open.page.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { AlertController } from '@ionic/angular';
 import { BaseProjectPage } from '../base-project.page';
 import { Project } from '../project';
 import { EventAggregatorService } from '../services/event-aggregator.service';
@@ -20,6 +21,7 @@ export class OpenPage extends BaseProjectPage implements OnInit {
     projectService: ProjectService,
     private events: EventAggregatorService,
     private router: Router,
+    private alertController: AlertController,
     private templateService: TemplateService) {
 
     super(route, projectService);
@@ -58,14 +60,53 @@ export class OpenPage extends BaseProjectPage implements OnInit {
     this.events.publish('CHANGE-MENU', 'req');
   }
 
-  async removeProject(prj: Project, event) {
+  async removeProject(prj: Project) {
+
+    if (this.projectId === prj.projectId) {
+      return;
+    }
+
     await this.projectService.removeProject(prj.projectId);
     const idx = this.projects.indexOf(prj);
     this.projects.splice(idx, 1);
+  }
+
+  async confirmRemove(prj: Project, event) {
+
     if (event) {
       event.preventDefault();
       event.cancelBubble = true;
     }
+
+    const alert = await this.alertController.create({
+      cssClass: 'alert-custom-class',
+      header: `Remove project '${prj.name}'?`,
+      message: 'If you want to access this project later, make sure to have saved it to disk before remove it from memory.',
+      buttons: [
+        {
+          text: 'Remove',
+          cssClass: 'danger',
+          handler: () => {
+            this.removeProject(prj);
+            return true;
+          }
+        },
+        {
+          text: 'Save it',
+          handler: () => {
+            this.templateService.exportToJson(this.projectId);
+            return true;
+          }
+        },
+        {
+          text: 'Cancel',
+          role: 'cancel',
+          cssClass: 'medium'
+        }
+      ]
+    });
+
+    await alert.present();
   }
 
 }
